Fix unauthed path handling.

+++ pages/new-room.tsx
@@ -3,6 +3,7 @@ import * as errors from '../frontend/lib/errors';
 import * as rooms from '../frontend/lib/rooms';
 import Loading from '../frontend/components/Loading';
 import { useRouter } from 'next/router';
+import firebase from 'firebase/app';
 
 let TeacherPage: React.FC = () => {
   let router = useRouter();
@@ -10,6 +11,10 @@ let TeacherPage: React.FC = () => {
   useEffect(() => {
     (async () => {
       try {
+        if (!firebase.auth().currentUser?.uid) {
+          return;
+        }
+
         await rooms.deleteCurrentRoom();
         let roomId = await rooms.generateId();
         router.push('/t/' + roomId);
